backend/requirements.txt
Flask
pandas
numpy
openpyxl
Pillow
Flask-Cors
cryptography
